Added multi scanning

+++ adhrit/recons/apk_extract.py
@@ -13,7 +13,7 @@ from adhrit.recons.enjarify.mutf8 import decode
 
 
 # To know the contents of a package
-def apk_info(apk_name):
+def apk_info(apk_name, hash_of_apk):
 
 # 	nlc = 0
 # 	apk = zipfile.ZipFile(apk_name, 'r')
@@ -114,21 +114,45 @@ def apk_info(apk_name):
 
 	print(Fore.YELLOW + "\n--------------------------------------------------")
 	print(Fore.GREEN + "[INFO] " + Fore.BLUE + "MANIFEST DUMP")
-	# os.chdir('..')
-	# os.system('rm AndroidManifest.xml')
-	# print(os.system('pwd'))
-	mandmp = 'java -jar adhrit/tools/apktool.jar d -f app.apk -o manifest'
-	os.system(mandmp)
-	os.system('mkdir api_scan')
-	if os.path.isdir('manifest'):
-		os.system('cp manifest/AndroidManifest.xml Manifest.xml')
-		os.system('cp manifest/AndroidManifest.xml api_scan/manifest.xml')
-		os.system('cp manifest/res/values/strings.xml api_scan/strings.xml')
-		os.system('cp -a manifest/smali api_scan/smali')
+	pwd = os.getcwd()
+	path = str(pwd) + '/'+hash_of_apk
+	try:
+		if not os.path.exists(path):
+			os.makedirs(hash_of_apk)
+			print('dir created')
+			os.chdir(path)
+			mandmp = 'java -jar ../adhrit/tools/apktool.jar d -sf ../app.apk -o ./manifest'
+			os.system(mandmp)
+			if os.path.isdir('manifest'):
+				os.system('cp manifest/AndroidManifest.xml Manifest.xml')
+				os.system('cp manifest/AndroidManifest.xml api_scan/Manifest.xml')
+				if os.path.isfile('manifest/res/values/strings.xml'):
+					os.system('cp manifest/res/values/strings.xml strings.xml')
+					os.system('cp manifest/res/values/strings.xml api_scan/strings.xml')
+			os.system('rm -rf manifest/')
+			path = path + '/..'
+			os.chdir(path)
+			
+	except OSError as err:
+		print('dir exist')
 
-		
+		cmd ='rm -rf '+ hash_of_apk
+		os.system(cmd)
 
+		os.makedirs(hash_of_apk)
+		print('dir created')
+		os.chdir(path)
+		mandmp = 'java -jar ../adhrit/tools/apktool.jar d -sf ../app.apk -o ./manifest'
+		os.system(mandmp)
+		if os.path.isdir('manifest'):
+			os.system('cp manifest/AndroidManifest.xml Manifest.xml')
+			os.system('cp manifest/res/values/strings.xml strings.xml')	
 		os.system('rm -rf manifest/')
+		path = path + '/..'
+		os.chdir(path)
+
+
+	
 		
 
 # 	# os.system('rm AndroidManifest.xml')
